\name{Rpu-package}
\alias{Rpu-package}
\alias{Rpu}
\docType{package}
\title{
\packageTitle{Rpu}
}
\description{
\packageDescription{Rpu}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rpu}
\packageIndices{Rpu}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Rpu}

Maintainer: \packageMaintainer{Rpu}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
